+++ AffindaAPI/AffindaAPI/AffindaAPIModelFactory.cs
@@ -102,37 +102,6 @@ namespace Affinda.API.Models
             return new User(id, name, username, email, avatar);
         }
 
-        /// <summary> Initializes a new instance of RequestError. </summary>
-        /// <param name="type"></param>
-        /// <param name="errors"></param>
-        /// <returns> A new <see cref="Models.RequestError"/> instance for mocking. </returns>
-        public static RequestError RequestError(string type = null, IEnumerable<RequestErrorErrorsItem> errors = null)
-        {
-            errors ??= new List<RequestErrorErrorsItem>();
-
-            return new RequestError(type, errors?.ToList());
-        }
-
-        /// <summary> Initializes a new instance of RequestErrorErrorsItem. </summary>
-        /// <param name="attr"></param>
-        /// <param name="code"></param>
-        /// <param name="detail"></param>
-        /// <exception cref="ArgumentNullException"> <paramref name="code"/> or <paramref name="detail"/> is null. </exception>
-        /// <returns> A new <see cref="Models.RequestErrorErrorsItem"/> instance for mocking. </returns>
-        public static RequestErrorErrorsItem RequestErrorErrorsItem(string attr = null, string code = null, string detail = null)
-        {
-            if (code == null)
-            {
-                throw new ArgumentNullException(nameof(code));
-            }
-            if (detail == null)
-            {
-                throw new ArgumentNullException(nameof(detail));
-            }
-
-            return new RequestErrorErrorsItem(attr, code, detail);
-        }
-
         /// <summary> Initializes a new instance of PathsZ1JuagV3WorkspaceMembershipsGetResponses200ContentApplicationJsonSchema. </summary>
         /// <param name="count"> Number of items in results. </param>
         /// <param name="next"> URL to request next page of results. </param>