+++ AffindaAPI/AffindaAPI/Models/JobDescription.cs
@@ -26,17 +26,13 @@ namespace Affinda.API.Models
         }
 
         /// <summary> Initializes a new instance of JobDescription. </summary>
+        /// <param name="data"> Any object. </param>
         /// <param name="extractor"></param>
         /// <param name="meta"></param>
         /// <param name="error"></param>
-        /// <param name="data"></param>
-        internal JobDescription(string extractor, DocumentMeta meta, DocumentError error, JobDescriptionData data) : base(extractor, meta, error)
+        internal JobDescription(object data, string extractor, DocumentMeta meta, DocumentError error) : base(data, extractor, meta, error)
         {
-            Data = data;
             Extractor = extractor ?? "job-description";
         }
-
-        /// <summary> Gets or sets the data. </summary>
-        public JobDescriptionData Data { get; set; }
     }
 }