+++ AffindaAPI/AffindaAPI/Models/RequestErrorErrorsItem.cs
@@ -10,7 +10,7 @@ using System;
 namespace Affinda.API.Models
 {
     /// <summary> The RequestErrorErrorsItem. </summary>
-    public partial class RequestErrorErrorsItem
+    internal partial class RequestErrorErrorsItem
     {
         /// <summary> Initializes a new instance of RequestErrorErrorsItem. </summary>
         /// <param name="attr"></param>