+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add AsInvoice(), AsResume() and AsJobDesription() methods to Document
+
+### Fixed
+- Document.Data is now present in return objects
 
 ## [5.4.0] - 2023-04-17
 ### Added